requirements.txt
google-generativeai
groq
PyPDF2
pytesseract
pdf2image